IPRT: Fixed -Wshadow warnings, found two bugs in error paths.
svn:sync-xref-src-repo-rev: r55308

+++ src/VBox/Runtime/testcase/tstTSC.cpp
@@ -1,4 +1,4 @@
-/* $Id: tstTSC.cpp 17093 2009-02-24 19:58:09Z knut.osmundsen@oracle.com $ */
+/* $Id: tstTSC.cpp 25000 2009-11-26 14:22:44Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT Testcase - SMP TSC testcase.
  */
@@ -389,7 +389,7 @@ static int tstTSCCalcFrequency(uint32_t cMsDuration)
      * Calc the frequency.
      */
     RTPrintf("tstTSC: %RU64 ticks in %RU64 ns\n", uTSC, uNanoTS);
-    uint64_t cHz = uTSC / ((long double)uNanoTS / (long double)1000000000);
+    uint64_t cHz = (uint64_t)(uTSC / ((long double)uNanoTS / (long double)1000000000));
     RTPrintf("tstTSC: Frequency %RU64 Hz", cHz);
     if (cHz > _1G)
     {
